Fix memory growing issue of the cached bLangPackages
Migrated all methods invoking LSPackageCache to load and define built-in
packages. Each invoke of the PackageLoader.loadAndDefine() will
re-populate the symbols such as functions..etc for that particular
package. Thus, moved the PackageLoader.loadAndDefine() as the back-off
method of the LSPackageLoader.getPackageById(...) method. Always use
LSPackageLoader to retrieve the packages. If it does not exists,
LSPackageLoader will call PackageLoader.loadAndDefine() from inside for
you :)