Add compensation support
This adds experimental support for the following syntax.
====================================================
scope scopeName {
          int varRef =5;
} compensation (varRef) { }
........
compensate scopeName;

=====================================================
There can be nested scopes as well.
"compensate" statement will execute the compensation blocks of the given
scope, including child scope compensations, in the reverse execution
order of the scopes. Variable references are passed to compensation
blocks with the value it had while completing the corresponding scope.

The variable passing part will be improved later. It has a lot of
limitations at the moment.