Set icons for the completion items
We can let the client(eg. VSCode) know the completion item kind to make
it more appealing and differentiate it between other items. With this
change we can have two distinguishable completion items that have
different icons for the error BType and Snippet Type. Latter type is for
the error declaration and first type is just for the keyword.

Signed-off-by: Rasika <info.rasika@gmail.com>