Avoid parsing JSON in setTextPayload
When the `contentType` header is set to `application/json` in `setTextPayload` function it tends to be serialized as JSON which is incorrect. This should be simply parsed as string. This PR fixes this issue. Whatever the header value is `setTextPayload` would parse it as string.

Resolve https://github.com/ballerina-platform/ballerina-lang/issues/15514